+++ verik-compiler/src/main/kotlin/io/verik/compiler/common/Visitor.kt
@@ -32,6 +32,7 @@ import io.verik.compiler.ast.element.declaration.common.EPackage
 import io.verik.compiler.ast.element.declaration.common.EProperty
 import io.verik.compiler.ast.element.declaration.common.ETypeParameter
 import io.verik.compiler.ast.element.declaration.kt.ECompanionObject
+import io.verik.compiler.ast.element.declaration.kt.EInitializerBlock
 import io.verik.compiler.ast.element.declaration.kt.EKtAbstractFunction
 import io.verik.compiler.ast.element.declaration.kt.EKtClass
 import io.verik.compiler.ast.element.declaration.kt.EKtFunction
@@ -233,6 +234,10 @@ abstract class Visitor {
 
     open fun visitSecondaryConstructor(secondaryConstructor: ESecondaryConstructor) {
         visitKtAbstractFunction(secondaryConstructor)
+    }
+
+    open fun visitInitializerBlock(initializerBlock: EInitializerBlock) {
+        visitAbstractFunction(initializerBlock)
     }
 
     open fun visitSvAbstractFunction(abstractFunction: ESvAbstractFunction) {